+++ COT/data_validation.py
@@ -46,7 +46,6 @@
   no_whitespace
   non_negative_int
   positive_int
-  to_string
   validate_int
   truth_value
 
@@ -56,40 +55,12 @@
   NIC_TYPES
 """
 
-import xml.etree.ElementTree as ET
 import hashlib
 import re
-import sys
+from collections import namedtuple
 from distutils.util import strtobool
 
-
-def to_string(obj):
-    """Get string representation of an object, special-case for XML Element.
-
-    Args:
-      obj (object): Object to represent as a string.
-    Returns:
-      str: string representation
-    Examples:
-      ::
-
-        >>> to_string("Hello")
-        'Hello'
-        >>> to_string(27.5)
-        '27.5'
-        >>> e = ET.Element('hello', attrib={'key': 'value'})
-        >>> print(e)   # doctest: +ELLIPSIS
-        <Element ...hello... at ...>
-        >>> print(to_string(e))
-        <hello key="value" />
-    """
-    if ET.iselement(obj):
-        if sys.version_info[0] >= 3:
-            return ET.tostring(obj, encoding='unicode')
-        else:
-            return ET.tostring(obj)
-    else:
-        return str(obj)
+from COT.utilities import to_string
 
 
 def alphanum_split(key):
@@ -261,7 +232,7 @@ def canonicalize_nic_subtype(subtype):
         Unsupported value 'foobar' for NIC subtype ...
 
     .. seealso::
-       :meth:`COT.platforms.GenericPlatform.validate_nic_type`
+       :meth:`COT.platforms.Platform.validate_nic_type`
     """
     return canonicalize_helper("NIC subtype", subtype,
                                _NIC_MAPPINGS, re.IGNORECASE)
@@ -473,7 +444,7 @@ def no_whitespace(string):
 
 def validate_int(string,
                  minimum=None, maximum=None,
-                 label="input"):
+                 label=None):
     """Parser helper function for validating integer arguments in a range.
 
     Args:
@@ -506,6 +477,8 @@ def validate_int(string,
         ...     print(e)
         Value '100' for x is too high - must be at most 10
     """
+    if label is None:
+        label = "input"
     try:
         i = int(string)
     except ValueError:
@@ -517,13 +490,14 @@ def validate_int(string,
     return i
 
 
-def non_negative_int(string):
+def non_negative_int(string, label=None):
     """Parser helper function for integer arguments that must be 0 or more.
 
     Alias for :func:`validate_int` setting :attr:`minimum` to 0.
 
     Args:
       string (str): String to validate.
+      label (str): Label to include in any errors raised
     Returns:
       int: Validated integer value
     Raises:
@@ -542,16 +516,17 @@ def non_negative_int(string):
         ...     print(e)
         Value '-1' for input is too low - must be at least 0
     """
-    return validate_int(string, minimum=0)
+    return validate_int(string, minimum=0, label=label)
 
 
-def positive_int(string):
+def positive_int(string, label=None):
     """Parser helper function for integer arguments that must be 1 or more.
 
     Alias for :func:`validate_int` setting :attr:`minimum` to 1.
 
     Args:
       string (str): String to validate.
+      label (str): Label to include in any errors raised
     Returns:
       int: Validated integer value
     Raises:
@@ -568,7 +543,7 @@ def positive_int(string):
         ...     print(e)
         Value '0' for input is too low - must be at least 1
     """
-    return validate_int(string, minimum=1)
+    return validate_int(string, minimum=1, label=label)
 
 
 def truth_value(value):
@@ -610,6 +585,10 @@ def truth_value(value):
             ['y', 'yes', 't', 'true', 'on', 1,
              'n', 'no', 'f', 'false', 'off', 0]
         )
+
+
+ValidRange = namedtuple('ValidRange', ['minimum', 'maximum'])
+"""Simple helper class representing a range of valid values."""
 
 
 # Some handy exception and error types we can throw
@@ -680,6 +659,6 @@ class ValueTooHighError(ValueUnsupportedError):
                         self.expected_value))
 
 
-if __name__ == "__main__":
+if __name__ == "__main__":   # pragma: no cover
     import doctest
     doctest.testmod()