+++ COT/helpers/helper.py
@@ -61,9 +61,6 @@ import distutils.spawn
 from distutils.version import StrictVersion
 import requests
 
-from verboselogs import VerboseLogger
-
-logging.setLoggerClass(VerboseLogger)
 logger = logging.getLogger(__name__)
 
 try:
@@ -708,6 +705,6 @@ def helper_select(choices):
     raise HelperNotFoundError(msg)
 
 
-if __name__ == "__main__":
+if __name__ == "__main__":   # pragma: no cover
     import doctest
     doctest.testmod()