+++ COT/platforms/tests/test_cisco_iosv.py
@@ -14,73 +14,69 @@
 
 """Unit test cases for IOSv platform."""
 
-import unittest
-import logging
-from logging import NullHandler
-
 from COT.platforms.cisco_iosv import IOSv
 from COT.data_validation import (
     ValueUnsupportedError, ValueTooLowError, ValueTooHighError
 )
-
-logging.getLogger('COT').addHandler(NullHandler())
+from COT.platforms.tests import PlatformTests
 
 # pylint: disable=missing-type-doc,missing-param-doc
 
 
-class TestIOSv(unittest.TestCase):
+class TestIOSv(PlatformTests.PlatformTest):
     """Test cases for Cisco IOSv platform handling."""
 
     cls = IOSv
+    product_string = "com.cisco.iosv"
 
     def test_nic_name(self):
         """Test NIC name construction."""
-        self.assertEqual(self.cls.guess_nic_name(1),
+        self.assertEqual(self.ins.guess_nic_name(1),
                          "GigabitEthernet0/0")
-        self.assertEqual(self.cls.guess_nic_name(2),
+        self.assertEqual(self.ins.guess_nic_name(2),
                          "GigabitEthernet0/1")
-        self.assertEqual(self.cls.guess_nic_name(3),
+        self.assertEqual(self.ins.guess_nic_name(3),
                          "GigabitEthernet0/2")
-        self.assertEqual(self.cls.guess_nic_name(4),
+        self.assertEqual(self.ins.guess_nic_name(4),
                          "GigabitEthernet0/3")
 
     def test_cpu_count(self):
         """Test CPU count limits."""
-        self.assertRaises(ValueTooLowError, self.cls.validate_cpu_count, 0)
-        self.cls.validate_cpu_count(1)
-        self.assertRaises(ValueTooHighError, self.cls.validate_cpu_count, 2)
+        self.assertRaises(ValueTooLowError, self.ins.validate_cpu_count, 0)
+        self.ins.validate_cpu_count(1)
+        self.assertRaises(ValueTooHighError, self.ins.validate_cpu_count, 2)
 
     def test_memory_amount(self):
         """Test RAM allocation limits."""
         self.assertRaises(ValueTooLowError,
-                          self.cls.validate_memory_amount, 191)
-        self.cls.validate_memory_amount(192)
-        self.cls.validate_memory_amount(3072)
+                          self.ins.validate_memory_amount, 191)
+        self.ins.validate_memory_amount(192)
+        self.ins.validate_memory_amount(3072)
         self.assertRaises(ValueTooHighError,
-                          self.cls.validate_memory_amount, 3073)
+                          self.ins.validate_memory_amount, 3073)
 
     def test_nic_count(self):
         """Test NIC range limits."""
-        self.assertRaises(ValueTooLowError, self.cls.validate_nic_count, -1)
-        self.cls.validate_nic_count(0)
-        self.cls.validate_nic_count(16)
-        self.assertRaises(ValueTooHighError, self.cls.validate_nic_count, 17)
+        self.assertRaises(ValueTooLowError, self.ins.validate_nic_count, -1)
+        self.ins.validate_nic_count(0)
+        self.ins.validate_nic_count(16)
+        self.assertRaises(ValueTooHighError, self.ins.validate_nic_count, 17)
 
     def test_nic_type(self):
         """Test NIC valid and invalid types."""
         self.assertRaises(ValueUnsupportedError,
-                          self.cls.validate_nic_type, "E1000e")
-        self.cls.validate_nic_type("E1000")
+                          self.ins.validate_nic_type, "E1000e")
+        self.ins.validate_nic_type("E1000")
         self.assertRaises(ValueUnsupportedError,
-                          self.cls.validate_nic_type, "PCNet32")
+                          self.ins.validate_nic_type, "PCNet32")
         self.assertRaises(ValueUnsupportedError,
-                          self.cls.validate_nic_type, "virtio")
+                          self.ins.validate_nic_type, "virtio")
         self.assertRaises(ValueUnsupportedError,
-                          self.cls.validate_nic_type, "VMXNET3")
+                          self.ins.validate_nic_type, "VMXNET3")
 
     def test_serial_count(self):
         """Test serial port range limits."""
-        self.assertRaises(ValueTooLowError, self.cls.validate_serial_count, 0)
-        self.cls.validate_serial_count(1)
-        self.cls.validate_serial_count(2)
-        self.assertRaises(ValueTooHighError, self.cls.validate_serial_count, 3)
+        self.assertRaises(ValueTooLowError, self.ins.validate_serial_count, 0)
+        self.ins.validate_serial_count(1)
+        self.ins.validate_serial_count(2)
+        self.assertRaises(ValueTooHighError, self.ins.validate_serial_count, 3)